+++ src/metrics/distance/hamming.js
@@ -17,9 +17,9 @@
 /**
  * Function returning the Hamming distance between two sequences.
  *
- * @param  {mixed}  a     - The first sequence to process.
- * @param  {mixed}  b     - The second sequence to process.
- * @return {number}       - The Hamming distance between a & b.
+ * @param  {mixed}  a - The first sequence to process.
+ * @param  {mixed}  b - The second sequence to process.
+ * @return {number}   - The Hamming distance between a & b.
  *
  * @throws {Error} The function expects sequences of equal length.
  */
@@ -44,9 +44,9 @@ export default function hamming(a, b) {
 /**
  * Function returning the normalized Hamming distance between two sequences.
  *
- * @param  {mixed}  a     - The first sequence to process.
- * @param  {mixed}  b     - The second sequence to process.
- * @return {number}       - The normalized Hamming distance between a & b.
+ * @param  {mixed}  a - The first sequence to process.
+ * @param  {mixed}  b - The second sequence to process.
+ * @return {number}   - The normalized Hamming distance between a & b.
  */
 export function normalizedDistance(a, b) {
 
@@ -69,9 +69,9 @@ export function normalizedDistance(a, b) {
 /**
  * Function returning the normalized Hamming similarity between two sequences.
  *
- * @param  {mixed}  a     - The first sequence to process.
- * @param  {mixed}  b     - The second sequence to process.
- * @return {number}       - The normalized Hamming similarity between a & b.
+ * @param  {mixed}  a - The first sequence to process.
+ * @param  {mixed}  b - The second sequence to process.
+ * @return {number}   - The normalized Hamming similarity between a & b.
  */
 export function normalizedSimilarity(a, b) {
   return 1 - normalizedDistance(a, b);
@@ -87,9 +87,9 @@ export function normalizedSimilarity(a, b) {
  * since JavaScript does not have a direct access to processor low-level ops
  * such as popcount, this should be the most performant we can do now.
  *
- * @param  {mixed}  a     - The first number to process.
- * @param  {mixed}  b     - The second number to process.
- * @return {number}       - The Hamming distance between a & b.
+ * @param  {mixed}  a - The first number to process.
+ * @param  {mixed}  b - The second number to process.
+ * @return {number}   - The Hamming distance between a & b.
  */
 export function bitwise(a, b) {
   let d = 0,